requirements.txt
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0